connectSqlDialog is an abstract class. Separated out Miaa and Postgress connection
