+++ extensions/arc/src/ui/dialogs/connectSqlDialog.ts
@@ -4,31 +4,26 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as azdata from 'azdata';
-import * as vscode from 'vscode';
 import { Deferred } from '../../common/promise';
-import { createCredentialId } from '../../common/utils';
-import { credentialNamespace } from '../../constants';
 import * as loc from '../../localizedConstants';
-import { ControllerModel } from '../../models/controllerModel';
-import { MiaaModel } from '../../models/miaaModel';
 import { InitializingComponent } from '../components/initializingComponent';
 
-export class ConnectToSqlDialog extends InitializingComponent {
-	private modelBuilder!: azdata.ModelBuilder;
+export abstract class ConnectToSqlDialog extends InitializingComponent {
+	protected modelBuilder!: azdata.ModelBuilder;
 
-	private serverNameInputBox!: azdata.InputBoxComponent;
-	private usernameInputBox!: azdata.InputBoxComponent;
-	private passwordInputBox!: azdata.InputBoxComponent;
-	private rememberPwCheckBox!: azdata.CheckBoxComponent;
+	protected serverNameInputBox!: azdata.InputBoxComponent;
+	protected usernameInputBox!: azdata.InputBoxComponent;
+	protected passwordInputBox!: azdata.InputBoxComponent;
+	protected rememberPwCheckBox!: azdata.CheckBoxComponent;
 
-	private _completionPromise = new Deferred<azdata.IConnectionProfile | undefined>();
+	protected _completionPromise = new Deferred<azdata.IConnectionProfile | undefined>();
 
-	constructor(private _controllerModel: ControllerModel, private _miaaModel: MiaaModel) {
+	constructor() {
 		super();
 	}
 
-	public showDialog(connectionProfile?: azdata.IConnectionProfile): azdata.window.Dialog {
-		const dialog = azdata.window.createModelViewDialog(loc.connectToSql(this._miaaModel.info.name));
+	public showDialog(instanceName: string, connectionProfile?: azdata.IConnectionProfile): azdata.window.Dialog {
+		const dialog = azdata.window.createModelViewDialog(loc.connectToSql(instanceName));
 		dialog.cancelButton.onClick(() => this.handleCancel());
 		dialog.registerContent(async view => {
 			this.modelBuilder = view.modelBuilder;
@@ -88,42 +83,7 @@ export class ConnectToSqlDialog extends InitializingComponent {
 		return dialog;
 	}
 
-	public async validate(): Promise<boolean> {
-		if (!this.serverNameInputBox.value || !this.usernameInputBox.value || !this.passwordInputBox.value) {
-			return false;
-		}
-		const connectionProfile: azdata.IConnectionProfile = {
-			serverName: this.serverNameInputBox.value,
-			databaseName: '',
-			authenticationType: 'SqlLogin',
-			providerName: 'MSSQL',
-			connectionName: '',
-			userName: this.usernameInputBox.value,
-			password: this.passwordInputBox.value,
-			savePassword: !!this.rememberPwCheckBox.checked,
-			groupFullName: undefined,
-			saveProfile: true,
-			id: '',
-			groupId: undefined,
-			options: {}
-		};
-		const result = await azdata.connection.connect(connectionProfile, false, false);
-		if (result.connected) {
-			connectionProfile.id = result.connectionId;
-			const credentialProvider = await azdata.credentials.getProvider(credentialNamespace);
-			if (connectionProfile.savePassword) {
-				await credentialProvider.saveCredential(createCredentialId(this._controllerModel.info.id, this._miaaModel.info.resourceType, this._miaaModel.info.name), connectionProfile.password);
-			} else {
-				await credentialProvider.deleteCredential(createCredentialId(this._controllerModel.info.id, this._miaaModel.info.resourceType, this._miaaModel.info.name));
-			}
-			this._completionPromise.resolve(connectionProfile);
-			return true;
-		}
-		else {
-			vscode.window.showErrorMessage(loc.connectToSqlFailed(this.serverNameInputBox.value, result.errorMessage));
-			return false;
-		}
-	}
+	public async abstract validate(): Promise<boolean>;
 
 	private handleCancel(): void {
 		this._completionPromise.resolve(undefined);
